COMP9900/capstoneproject-comp9900-h18a-yyds/YYDSFinalSoftwareQuality/backend/requirements.txt
dnspython==2.1.0
passlib==1.7.4
Flask==1.1.4 
Jinja2==2.11.3 
click==7.1.2 
flask-restx==0.4.0 
itsdangerous==1.1.0 
werkzeug==1.0.1
flask-cors==3.0.10
pymongo==3.11.4
Pillow==8.2.0
PyJWT==2.1.0 
flask-jwt-extended==4.2.1
certifi==2021.5.30
flask-caching==1.10.1
redis==3.5.3
numpy==1.21.1
blinker==1.4 
flask-mail==0.9.1